% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_list_globally}
\alias{assign_list_globally}
\title{Helper function to assign list elements into global environment}
\usage{
assign_list_globally(x)
}
\arguments{
\item{x}{list object.}
}
\value{
The global environment invisibly.
}
\description{
This function is a wrapper around \code{list2env} that assigns the every object in
the list to the global environment as individual objects.

This function is useful for loading dimension tables into the global
environment as individual dataframes or breaking down the prelude and pulling
it into the global environment for inspection.
}
\examples{
\dontrun{
mylist <- dplyr::lst(
  table1 = mtcars,
  table2 = mtcars,
  table3 = mtcars
)

assign_list_globally(mylist)
}
}
