% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{nuke_table}
\alias{nuke_table}
\title{Drop a SQL Table}
\usage{
nuke_table(conn, database, schema, table)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{database}{Name of database or catalog.}

\item{schema}{Name of schema.}

\item{table}{Name of table.}
}
\value{
TRUE invisibly.
}
\description{
Drop a SQL Table
}
\examples{
\dontrun{
nuke_table(
  conn = my_connection,
  database = "my_database",
  schema = "my_schema",
  table = "my_table"
)
}
}
\seealso{
Other basics: 
\code{\link{append_table}()},
\code{\link{create_table}()},
\code{\link{overwrite_table}()},
\code{\link{read_table}()},
\code{\link{survey_table}()},
\code{\link{truncate_table}()}
}
\concept{basics}
