% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{overwrite_table}
\alias{overwrite_table}
\title{Truncate and Append a SQL table - Preserving Structure}
\usage{
overwrite_table(.data, conn, database, schema, table)
}
\arguments{
\item{.data}{Dataframe to append to the SQL table.}

\item{conn}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{database}{Name of database or catalog.}

\item{schema}{Name of schema.}

\item{table}{Name of table.}
}
\value{
A scalar numeric.
}
\description{
Truncate and append data to a SQL table without losing
SQL identities and without changing the table structure.
}
\examples{
\dontrun{
overwrite_table(
  .data = my_dataframe,
  conn = my_connection,
  database = "my_database",
  schema = "my_schema",
  table = "my_table"
)
}
}
\seealso{
Other basics: 
\code{\link{append_table}()},
\code{\link{create_table}()},
\code{\link{nuke_table}()},
\code{\link{read_table}()},
\code{\link{survey_table}()},
\code{\link{truncate_table}()}
}
\concept{basics}
