% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{read_table}
\alias{read_table}
\title{Read all data from a SQL Table}
\usage{
read_table(conn, database, schema, table, lazy = FALSE)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{database}{Name of database or catalog.}

\item{schema}{Name of schema.}

\item{table}{Name of table.}

\item{lazy}{Bool with default value \code{FALSE} indicating a tibble in memory
will be returned to the user. \code{TRUE} specifies that a lazy tibble will be
returned.}
}
\value{
Tibble or lazy tibble.
}
\description{
Read all rows of data from a SQL table as either a dataframe
stored in memory or as a lazy tibble.
}
\examples{
\dontrun{
read_table(
  conn = my_connection,
  database = "my_database",
  schema = "my_schema",
  table = "my_table",
  lazy = FALSE
)
}
}
\seealso{
Other basics: 
\code{\link{append_table}()},
\code{\link{create_table}()},
\code{\link{nuke_table}()},
\code{\link{overwrite_table}()},
\code{\link{survey_table}()},
\code{\link{truncate_table}()}
}
\concept{basics}
