% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{survey_table}
\alias{survey_table}
\title{Survey information about a SQL table}
\usage{
survey_table(conn, database, schema, table)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{database}{Name of database or catalog.}

\item{schema}{Name of schema.}

\item{table}{Name of table.}
}
\value{
Tibble with information about the target table.
}
\description{
\code{survey_table} returns a tibble with information about the target SQL table.
}
\examples{
\dontrun{
survey_table(
  conn = my_connection,
  database = "my_database",
  schema = "my_schema",
  table = "my_table"
)
}
}
\seealso{
Other basics: 
\code{\link{append_table}()},
\code{\link{create_table}()},
\code{\link{nuke_table}()},
\code{\link{overwrite_table}()},
\code{\link{read_table}()},
\code{\link{truncate_table}()}
}
\concept{basics}
