% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{truncate_table}
\alias{truncate_table}
\title{Truncate a SQL table}
\usage{
truncate_table(conn, database, schema, table)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{database}{Name of database or catalog.}

\item{schema}{Name of schema.}

\item{table}{Name of table.}
}
\value{
A scalar numeric that specifies the number of rows affected by the
truncate command.
}
\description{
\code{truncate_table} removes all rows from a SQL table but leaves the table
structure and identity intact.
}
\examples{
\dontrun{
truncate_table(
  conn = my_connection,
  database = "my_database",
  schema = "my_schema",
  table = "my_table"
)
}
}
\seealso{
Other basics: 
\code{\link{append_table}()},
\code{\link{create_table}()},
\code{\link{nuke_table}()},
\code{\link{overwrite_table}()},
\code{\link{read_table}()},
\code{\link{survey_table}()}
}
\concept{basics}
